created window.rs to handle build_ui function.
